Break the massive api/admin/controller __init__.py file into different files.
