+++ api/admin/controller/sitewide_settings.py
@@ -2,11 +2,13 @@ import flask
 from flask import Response
 from flask_babel import lazy_gettext as _
 
-from api.admin.problem_details import *
+from api.admin.controller.settings import SettingsController
+from api.admin.problem_details import (
+    MISSING_SITEWIDE_SETTING_KEY,
+    MISSING_SITEWIDE_SETTING_VALUE,
+)
 from api.config import Configuration
 from core.model import ConfigurationSetting
-
-from . import SettingsController
 
 
 class SitewideConfigurationSettingsController(SettingsController):
