+++ api/admin/controller/storage_services.py
@@ -1,14 +1,13 @@
 import flask
 from flask import Response
 
-from api.admin.problem_details import *
+from api.admin.controller.settings import SettingsController
+from api.admin.problem_details import CANNOT_CHANGE_PROTOCOL, MISSING_SERVICE
 
 # NOTE: We need to import it explicitly to initialize MirrorUploader.IMPLEMENTATION_REGISTRY
 from core.mirror import MirrorUploader  # noqa: autoflake
 from core.model import ExternalIntegration, get_one
 from core.util.problem_detail import ProblemDetail
-
-from . import SettingsController
 
 
 class StorageServicesController(SettingsController):
