+++ api/controller.py
@@ -82,7 +82,7 @@ from core.util.datetime_helpers import utc_now
 from core.util.http import HTTP, RemoteIntegrationException
 from core.util.log import elapsed_time_logging, log_elapsed_time
 from core.util.opds_writer import OPDSFeed
-from core.util.problem_detail import ProblemDetail, ProblemError
+from core.util.problem_detail import ProblemError
 from core.util.string_helpers import base64
 
 from .annotations import AnnotationParser, AnnotationWriter
@@ -122,25 +122,40 @@ from .shared_collection import SharedCollectionAPI
 if TYPE_CHECKING:
     from werkzeug import Response as wkResponse
 
+    from api.admin.controller.admin_search import AdminSearchController
+    from api.admin.controller.analytics_services import AnalyticsServicesController
     from api.admin.controller.announcement_service import AnnouncementSettings
     from api.admin.controller.catalog_services import CatalogServicesController
     from api.admin.controller.collection_self_tests import CollectionSelfTestsController
     from api.admin.controller.collection_settings import CollectionSettingsController
+    from api.admin.controller.custom_lists import CustomListsController
+    from api.admin.controller.dashboard_controller import DashboardController
+    from api.admin.controller.discovery_service_library_registrations import (
+        DiscoveryServiceLibraryRegistrationsController,
+    )
+    from api.admin.controller.discovery_services import DiscoveryServicesController
+    from api.admin.controller.feed import FeedController
     from api.admin.controller.individual_admin_settings import (
         IndividualAdminSettingsController,
     )
+    from api.admin.controller.lanes import LanesController
     from api.admin.controller.library_settings import LibrarySettingsController
     from api.admin.controller.metadata_service_self_tests import (
         MetadataServiceSelfTestsController,
     )
     from api.admin.controller.metadata_services import MetadataServicesController
+    from api.admin.controller.patron import PatronController
     from api.admin.controller.patron_auth_service_self_tests import (
         PatronAuthServiceSelfTestsController,
     )
     from api.admin.controller.patron_auth_services import PatronAuthServicesController
+    from api.admin.controller.reset_password import ResetPasswordController
     from api.admin.controller.search_service_self_tests import (
         SearchServiceSelfTestsController,
     )
+    from api.admin.controller.self_tests import SelfTestsController
+    from api.admin.controller.settings import SettingsController
+    from api.admin.controller.sign_in import SignInController
     from api.admin.controller.sitewide_services import (
         LoggingServicesController,
         SearchServicesController,
@@ -150,27 +165,9 @@ if TYPE_CHECKING:
         SitewideConfigurationSettingsController,
     )
     from api.admin.controller.storage_services import StorageServicesController
+    from api.admin.controller.timestamps import TimestampsController
+    from api.admin.controller.view import ViewController
     from api.admin.controller.work_editor import WorkController as AdminWorkController
-
-    from .admin.controller import (
-        AdminSearchController,
-        CustomListsController,
-        DashboardController,
-        FeedController,
-        LanesController,
-        PatronController,
-        ResetPasswordController,
-        SettingsController,
-        SignInController,
-        TimestampsController,
-        ViewController,
-    )
-    from .admin.controller.analytics_services import AnalyticsServicesController
-    from .admin.controller.discovery_service_library_registrations import (
-        DiscoveryServiceLibraryRegistrationsController,
-    )
-    from .admin.controller.discovery_services import DiscoveryServicesController
-    from .admin.controller.self_tests import SelfTestsController
 
 
 class CirculationManager:
